flashを追加した。全体のflash含めてDRYが可能か検討する.

+++ app/controllers/tags_controller.rb
@@ -3,7 +3,7 @@ class TagsController < ApplicationController
     @tag = @current_user.tag.new(tags_params)
     respond_to do |format|
       if @tag.save
-        format.js
+        format.js { flash.now[:success] = "タグを作成しました。" }
       else
         format.js { render :create_errors }
       end
@@ -15,7 +15,7 @@ class TagsController < ApplicationController
     @tag = @current_user.tag.find(tags_params[:id])
     respond_to do |format|
       if @tag.update(tags_params)
-        format.js
+        format.js { flash.now[:success] = "タグを更新しました。" }
       else
         format.js { render :edit_errors }
       end
@@ -24,16 +24,21 @@ class TagsController < ApplicationController
   end
 
   def destroy
-    select_tags = select_tags_params
-    respond_to do |format|
-      select_tags.each do |tag|
-        @tag = Tag.find(tag)
-        if @tag.destroy
-          format.js
-        else
-          format.js { render :edit_errors}
+    if select_tags_params.present?
+      select_tags = select_tags_params
+      respond_to do |format|
+        select_tags.each do |tag|
+          @tag = Tag.find(tag)
+          if @tag.destroy
+            format.js { flash.now[:success] = "タグを削除しました。" }
+          else
+            format.js { render template: "users/show" }
+            flash.now[:alert] = "タグを削除できませんでした。"
+          end
         end
       end
+    else
+      flash.now[:success] = "タグを選択してください"
     end
     set_tags
   end
@@ -45,8 +50,8 @@ class TagsController < ApplicationController
   end
 
   def select_tags_params
-  ids = params.require(:tag).permit(tags_ids: [])
-  ids.values[0]
+    ids = params.fetch(:tag, {}).permit(tags_ids: [])
+    ids.values[0]
   end
 
   def set_tags
@@ -54,5 +59,4 @@ class TagsController < ApplicationController
     @can_tags = @current_user.tag.where(wcm: "can")
     @must_tags = @current_user.tag.where(wcm: "must")
   end
-
 end
